requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.2.2
South==0.8.4
distribute==0.6.10
django-admin-shortcuts==1.2.5
django-custom-user==0.2
django-mptt==0.6.1
django-nose==1.2
django-rest-utils==0.0.5
djangocms-admin-style==0.2.2
djangorestframework==2.4.3
nose==1.3.4

